Conflicts:
	CloverInWpf/Clover/AbstractLayer/CloverController.cs
	CloverInWpf/Clover/AbstractLayer/FaceLayer.cs
	CloverInWpf/Clover/AbstractLayer/FoldingSystem.cs

+++ CloverInWpf/Clover/AbstractLayer/VertexLayer.cs
@@ -8,6 +8,7 @@ namespace Clover
     /// <summary>
     /// Basic vertex structure.
     /// </summary>
+    [Serializable]
     public class VertexLayer
     {
         #region Attributes
@@ -21,7 +22,6 @@ namespace Clover
 
         // Vertexcell lookup table.
         List<List<Vertex>> vertexCellTable = new List<List<Vertex>>();
-        CloverController controller;
 
         #endregion
 
@@ -43,9 +43,8 @@ namespace Clover
         #region Public Interfaces.
 
         // Constructor and Destroyer.
-        public VertexLayer(CloverController ctrl)
+        public VertexLayer()
         {
-            controller = ctrl;
         }
 
         public bool IsVertexExist(int index)
@@ -133,12 +132,27 @@ namespace Clover
         }
 
         /// <summary>
-        /// 将顶点删除到当前节点的版本
+        /// 将当前节点之后的版本删掉
+        /// </summary>
+        /// <param name="vertex"></param>
+        public void DeleteNextVersionToEnd(Vertex vertex)
+        {
+            int index = VertexCellTable[vertex.Index].IndexOf(vertex) + 1;
+            if (index != -1 && index < VertexCellTable[vertex.Index].Count)
+            {
+                VertexCellTable[vertex.Index].RemoveRange(index, VertexCellTable[vertex.Index].Count - index);
+            }
+        }
+
+        /// <summary>
+        /// 把当前节点和之后的节点都删除
         /// </summary>
         /// <param name="vertex"></param>
         public void DeleteThisVersionToEnd(Vertex vertex)
         {
             int index = VertexCellTable[vertex.Index].IndexOf(vertex);
+            if (index == -1)
+                return;
             VertexCellTable[vertex.Index].RemoveRange(index, VertexCellTable[vertex.Index].Count - index);
         }
